Started writing the Item abstract class
Wrote a few functions to add items to a check, and add/remove items from
the menu. I am thinking that the subclasses (i.e. drinks/entrees etc.)
will store the names of the items, and the view can get this information
when it creates the buttons. I did not modify anything in the view
though. If I am incorrect on this assumption let me know.