Editted the LoginModel
Before, the initialization for each server's view was when the PIN
matched. So, a new view/controller was being created each time the
server logged in. If our program was running for several hours, I could
definitely see that impacting performance. Changed the class so that all
views are initialized when program is launched. They are all just set to
not be visible unless the PIN is correct.